[IMP] outpostnosara: The room assignation now considers the datetimes too
Before this change the assignation of a preferred room was considering
just if the room is available the whole day, now the rooms can be
reserved considering hours too, for example, it is possible to
reserve a room from 8 am to 1 pm and after the same day from 2 pm to 6
pm.

The system gives priority to those rooms that are not being used that
day in order to avoid performance issues, maybe in the future this can
be improved.

Task#50570