[ADD] outpostnosara: Set required fields to l10n_cr invoice flow
Before this change customer was creating lots of products without CABYS
code neither unit of meature for eInvoicing. Additionall were presenting
to DGT several invoices with economic activity by default which is for
Coffee shop and not for leasing.

With this change those problems are fixed making required the
information for products, sale orders and subscriptions to prevent the
creation of invoices without correct configuration data.

Issue reported on task #50621