[ADD] l10n_cr_edi_discount_pos: New module welcome
Normally odoo sets a negative line when a global discount is
applied on a POS order and when that order is invoiced,
the negative line is kept in the Invoice. Depending the
implementation or the country, this behavior is not correct.

This module changes this behavior, to apply the discount proportionally in each invoice line.

Task#57817